Reverting puller
