+++ cylibs/trust/roles/puller.lua
@@ -2,17 +2,15 @@ local Approach = require('cylibs/battle/approach')
 local ClaimedCondition = require('cylibs/conditions/claimed')
 local ConditionalCondition = require('cylibs/conditions/conditional')
 local DisposeBag = require('cylibs/events/dispose_bag')
-local Distance = require('cylibs/conditions/distance')
 local Engage = require('cylibs/battle/engage')
 local Gambit = require('cylibs/gambits/gambit')
 local GambitTarget = require('cylibs/gambits/gambit_target')
 local MobFilter = require('cylibs/battle/monsters/mob_filter')
-local PartyClaimedCondition = require('cylibs/conditions/party_claimed')
 local PartyLeaderCondition = require('cylibs/conditions/party_leader')
 local PartyMemberCountCondition = require('cylibs/conditions/party_member_count')
 local PartyTargetedCondition = require('cylibs/conditions/party_targeted')
 local RunToLocationAction = require('cylibs/actions/runtolocation')
-local TargetLock = require('cylibs/entity/party/target_lock')
+local TargetMismatchCondition = require('cylibs/conditions/target_mismatch')
 local Timer = require('cylibs/util/timers/timer')
 
 local Gambiter = require('cylibs/trust/roles/gambiter')
@@ -39,7 +37,6 @@ function Puller.new(action_queue, pull_settings, job)
 
     self.job = job
     self.target_timer = Timer.scheduledTimer(1, 0)
-    self.assist_target_dispose_bag = DisposeBag.new()
     self.dispose_bag = DisposeBag.new()
     self.dispose_bag:addAny(L{ self.target_timer })
 
@@ -54,13 +51,6 @@ function Puller:destroy()
     self.dispose_bag:destroy()
 end
 
-function Puller:target_change(target_index)
-    Gambiter.target_change(self, target_index)
-
-    print('target_change', 'checking gambits')
-    self:check_gambits(nil, nil, true)
-end
-
 function Puller:on_add()
     Gambiter.on_add(self)
 
@@ -69,60 +59,44 @@ function Puller:on_add()
         self:set_pull_target(Monster.new(current_target.id))
     end
 
-    local on_pull_mode_changed = function(new_value)
-        if new_value ~= 'Off' then
-            self:set_pull_target(nil)
+    if state.AutoPullMode.value ~= 'Off' then
+        windower.send_command('input /autotarget off')
+    end
 
+    self.dispose_bag:add(state.AutoPullMode:on_state_change():addAction(function(_, new_value)
+        if new_value ~= 'Off' then
             windower.send_command('input /autotarget off')
             local assist_target = self:get_party():get_assist_target()
             if assist_target:get_id() ~= windower.ffxi.get_player().id then
+                self:get_party():add_to_chat(self:get_party():get_player(), "I can't pull while I'm assisting someone else, so I'm going to stop assisting "..assist_target:get_name()..".")
                 self:get_party():set_assist_target(self:get_party():get_player())
             end
-        else
-            self:set_pull_target(nil)
         end
-    end
-    on_pull_mode_changed(state.AutoPullMode.value)
-
-    self.dispose_bag:add(state.AutoPullMode:on_state_change():addAction(function(_, new_value)
-        on_pull_mode_changed(new_value)
     end), state.AutoPullMode:on_state_change())
 
     self.dispose_bag:add(self.target_timer:onTimeChange():addAction(function(_)
         if not addon_enabled:getValue() then
             return
         end
-        -- return to camp might be broken now that this doesn't fire as often
-        --self:check_target()
+        self:check_target()
     end, self.target_timer:onTimeChange()))
 
     self.dispose_bag:add(WindowerEvents.MobKO:addAction(function(mob_id, mob_name, status)
-        --if state.AutoPullMode.value == 'Off' then
-        --    return
-        --end
         if self:get_target() and self:get_target():get_id() == mob_id then
             logger.notice(self.__class, 'mob_ko', mob_name, self:get_target():get_mob().hpp, status)
-            print(os.time(), 'mob KO', 'picking new target')
-            --self:set_pull_target(nil)
+            self:set_pull_target(nil)
 
             self:check_target(L{ mob_id })
-
-            --coroutine.schedule(function()
-            --    if self:get_pull_target() then
-            --        print(os.time(), 'found target', self.target_lock ~= nil, self:get_all_gambits():length())
-            --        self:check_gambits(nil, nil, true)
-            --    end
-            --end, 0.05)
+            if self:get_pull_target() then
+                self:check_gambits(nil, nil, true)
+            end
         end
-
     end), WindowerEvents.MobKO)
 
     self.target_timer:start()
 end
 
-function Puller:tic(old_time, new_time)
-    Gambiter.tic(self, old_time, new_time)
-
+function Puller:tic(_, _)
     if state.AutoPullMode.value == 'Off' then
         return
     end
@@ -144,12 +118,11 @@ function Puller:check_target(target_id_blacklist)
             local previous_target = next_target:get_mob()
             logger.notice(self.__class, 'check_target', 'clear', previous_target.name, previous_target.hpp, previous_target.index, previous_target.status, previous_target.claim_id or 'unclaimed')
         end
-        print(os.time(), 'set to nil')
+
         self:set_pull_target(nil)
 
         next_target = self:get_next_target(target_id_blacklist)
         if next_target then
-            print(os.time(), 'set to non nil')
             self:set_pull_target(next_target)
             logger.notice(self.__class, 'check_target', 'set_pull_target', next_target:get_name(), next_target:get_mob().index)
         else
@@ -179,8 +152,8 @@ function Puller:get_all_targets()
         all_targets = self.mob_filter:get_aggroed_mobs(L{ MobFilter.Type.PartyClaimed })
                 + self.mob_filter:get_aggroed_mobs(L{ MobFilter.Type.Unclaimed })
                 + (self.mob_filter:get_nearby_mobs(L{ MobFilter.Type.Unclaimed }):filter(function(mob)
-                    return self.target_names:contains(mob.name)
-                end))
+            return self.target_names:contains(mob.name)
+        end))
     elseif state.AutoPullMode.value == 'All' then
         -- 1. All mobs that are party claimed
         -- 2. All mobs that are unclaimed
@@ -201,7 +174,7 @@ function Puller:get_next_target(target_id_blacklist)
         return not target_id_blacklist:contains(target.id) and self:is_valid_target(target)
     end)
     if all_targets:length() > 0 then
-        if state.PullActionMode.value == 'Target' or self.max_num_targets > 1 then
+        if state.PullActionMode.value == 'Target' or self.max_num_targets > 1 then -- TODO: test whether this randomizes target
             return Monster.new(all_targets:random().id)
         else
             return Monster.new(all_targets[1].id)
@@ -215,45 +188,40 @@ function Puller:is_valid_target(target)
     if not target then
         return false
     end
+
+    local max_pull_ability_range = 0
+    for gambit in self:get_pull_abilities():it() do
+        max_pull_ability_range = math.max(max_pull_ability_range, gambit:getAbility():get_range())
+    end
     local conditions = L{
         MinHitPointsPercentCondition.new(1),
         ConditionalCondition.new(L{
-            PartyClaimedCondition.new(true),
-            ConditionalCondition.new(L{ UnclaimedCondition.new(), MaxDistanceCondition.new(self.max_pull_ability_range[state.PullActionMode.value]) }, Condition.LogicalOperator.And)
+            ClaimedCondition.new(self:get_party():get_party_members(true):map(function(p) return p:get_id() end)), -- TODO: should probably be alliance member ids
+            ConditionalCondition.new(L{ UnclaimedCondition.new(), MaxDistanceCondition.new(max_pull_ability_range) }, Condition.LogicalOperator.And)
         }, Condition.LogicalOperator.Or),
     }
     return not L{ 2, 3 }:contains(target.status) and Condition.check_conditions(conditions, target.index)
 end
 
 function Puller:get_pull_target()
-    return self:get_target()
+    return self.target
 end
 
 function Puller:set_pull_target(target)
-    local target_lock = self:get_party():get_assist_target()
-    if target and target:get_id() == target_lock:get_id() then
+    if target and self.target
+            and target:get_id() == self.target:get_id() then
         return
     end
-    if self.target_lock == nil and target == nil then
-        return
+    if self.target then
+        self.target:destroy()
     end
-    self.target_lock = nil
-    self.assist_target_dispose_bag:dispose()
+    self.target = target
+end
 
-    local assist_target = self:get_party():get_player()
-    if state.AutoPullMode.value ~= 'Off' and target then
-        assist_target = TargetLock.new(target:get_mob().index, self:get_party())
-        assist_target:monitor()
-
-        self.assist_target_dispose_bag:addAny(L{ assist_target })
-
-        self.target_lock = assist_target
-    else
-        if windower.ffxi.get_player().locked_on then
-            windower.send_command('input /lockon')
-        end
-    end
-    self:get_party():set_assist_target(assist_target)
+function Puller:get_gambit_targets(gambit_target_types)
+    local targets_by_type = Gambiter.get_gambit_targets(self, gambit_target_types)
+    targets_by_type[GambitTarget.TargetType.Enemy] = L{ self:get_pull_target() }:compact_map()
+    return targets_by_type
 end
 
 function Puller:get_pull_settings()
@@ -274,50 +242,30 @@ function Puller:set_pull_settings(pull_settings)
         gambit.conditions = gambit.conditions:filter(function(condition)
             return condition:is_editable()
         end)
-        local conditions = self:get_default_conditions(gambit, 'PullActionMode', 'Auto')
+        local conditions = self:get_default_conditions(gambit)
         for condition in conditions:it() do
             condition:set_editable(false)
             gambit:addCondition(condition)
         end
     end
 
-    local approach = Gambit.new(GambitTarget.TargetType.Enemy, L{}, Approach.new(L{MaxDistanceCondition.new(35)}), GambitTarget.TargetType.Enemy, L{"Pulling"})
-    approach.conditions = self:get_default_conditions(approach, 'PullActionMode', 'Approach')
-
-    local auto_target = Gambit.new(GambitTarget.TargetType.Enemy, L{}, Engage.new(L{MaxDistanceCondition.new(30)}), GambitTarget.TargetType.Enemy, L{"Pulling","Reaction"})
-    auto_target.conditions = self:get_default_conditions(auto_target, 'PullActionMode', 'Target')
-
-    self.max_pull_ability_range = {
-        Auto = self:get_max_pull_ability_range(pull_settings.Gambits),
-        Approach = approach:getAbility():get_range(),
-        Target = auto_target:getAbility():get_range(),
-    }
-
-    local gambit_settings = {
-        Gambits = L{ approach } + L{ auto_target } + pull_settings.Gambits
-    }
-
     self:set_target_names(pull_settings.Targets or L{})
-    self:set_gambit_settings(gambit_settings)
 end
 
-function Puller:get_default_conditions(gambit, mode_name, mode_value)
+function Puller:get_default_conditions(gambit)
     local conditions = L{
-        GambitCondition.new(ModeCondition.new(mode_name, mode_value), GambitTarget.TargetType.Self),
-        GambitCondition.new(UnclaimedCondition.new(), GambitTarget.TargetType.Enemy),
         GambitCondition.new(MaxDistanceCondition.new(gambit:getAbility():get_range()), GambitTarget.TargetType.Enemy),
-        --GambitCondition.new(Distance.new(3, Condition.Operator.GreaterThanOrEqualTo), GambitTarget.TargetType.Enemy),
         GambitCondition.new(MinHitPointsPercentCondition.new(1), GambitTarget.TargetType.Enemy),
         GambitCondition.new(NotCondition.new(L{ HasBuffCondition.new('weakness') }), GambitTarget.TargetType.Self)
     }
     local alter_ego_conditions = L{
         GambitCondition.new(ConditionalCondition.new(
-            L{
-                NotCondition.new(L{ PartyLeaderCondition.new() }),
-                ModeCondition.new('AutoTrustsMode', 'Off'),
-                ConditionalCondition.new(L{ ModeCondition.new('AutoTrustsMode', 'Auto'), ModeCondition.new('AutoPullMode', 'Auto'), PartyMemberCountCondition.new(6, Condition.Operator.GreaterThanOrEqualTo) }, Condition.LogicalOperator.And)
-            },
-            Condition.LogicalOperator.Or), GambitTarget.TargetType.Self)
+                L{
+                    NotCondition.new(L{ PartyLeaderCondition.new() }),
+                    ModeCondition.new('AutoTrustsMode', 'Off'),
+                    ConditionalCondition.new(L{ ModeCondition.new('AutoTrustsMode', 'Auto'), ModeCondition.new('AutoPullMode', 'Auto'), PartyMemberCountCondition.new(6, Condition.Operator.GreaterThanOrEqualTo) }, Condition.LogicalOperator.And)
+                },
+                Condition.LogicalOperator.Or), GambitTarget.TargetType.Self)
     }
     return (alter_ego_conditions + conditions + self.job:get_conditions_for_ability(gambit:getAbility())):map(function(condition)
         if condition.__type ~= GambitCondition.__type then
@@ -327,28 +275,34 @@ function Puller:get_default_conditions(gambit, mode_name, mode_value)
     end)
 end
 
-function Puller:get_max_pull_ability_range(gambits)
-    local max_pull_ability_range = 0
-    for gambit in gambits:it() do
-        max_pull_ability_range = math.max(max_pull_ability_range, gambit:getAbility():get_range())
+function Puller:get_pull_abilities()
+    if state.PullActionMode.value == 'Approach' then
+        local approach = Gambit.new(GambitTarget.TargetType.Enemy, L{}, Approach.new(L{MaxDistanceCondition.new(35)}), GambitTarget.TargetType.Enemy, L{"Pulling"})
+        approach.conditions = self:get_default_conditions(approach)
+        return L{ approach }
+    elseif state.PullActionMode.value == 'Target' then
+        local auto_target = Gambit.new(GambitTarget.TargetType.Enemy, L{}, Engage.new(L{MaxDistanceCondition.new(30)}), GambitTarget.TargetType.Enemy, L{"Pulling","Reaction"})
+        auto_target.conditions = self:get_default_conditions(auto_target)
+        return L{ auto_target }
     end
-    return max_pull_ability_range
+    return self.pull_abilities
 end
 
 function Puller:get_all_gambits()
-    if self.target_lock == nil then
+    local next_target = self:get_pull_target()
+    if not next_target or not self:is_valid_target(next_target and next_target:get_mob()) then
         return L{}
     end
-    return Gambiter.get_all_gambits(self)
-end
 
-function Puller:get_gambit_targets(gambit_target_types)
-    local targets_by_type = Gambiter.get_gambit_targets(self, gambit_target_types)
-    if self.target_lock then
-        print('target lock')
-        targets_by_type[GambitTarget.TargetType.Enemy] = L{ Monster.new(self.target_lock.target_id) }:compact_map()
+    if next_target:is_unclaimed() then
+        return self:get_pull_abilities()
+    elseif next_target:is_claimed_by(self:get_alliance()) and (self:get_target() ~= next_target or self:get_target() == next_target and self:get_party():get_player():get_status() ~= 'Engaged') then
+        local auto_target = Gambit.new(GambitTarget.TargetType.Enemy, L{}, Engage.new(L{MaxDistanceCondition.new(30)}, true), GambitTarget.TargetType.Enemy, L{"Pulling"})
+        auto_target.conditions = self:get_default_conditions(auto_target)
+        return L{ auto_target }
     end
-    return targets_by_type
+
+    return L{}
 end
 
 function Puller:get_type()
@@ -372,7 +326,7 @@ function Puller:get_camp_position()
 end
 
 function Puller:get_cooldown()
-    return 3
+    return 5
 end
 
 function Puller:allows_duplicates()
@@ -384,9 +338,8 @@ function Puller:allows_multiple_actions()
 end
 
 function Puller:return_to_camp()
-    if state.AutoCampMode.value == 'Off' or self:get_camp_position() == nil or self:get_party():get_player():get_status() == 'Engaged'
-            or self:get_target() ~= nil or self:get_party():get_assist_target().__type == TargetLock.__type then
-        --print('cant return to camp', self:get_pull_target() ~= nil, self:get_party():get_assist_target().__type == TargetLock.__type, self:get_camp_position() == nil)
+    if state.AutoCampMode.value == 'Off' or self:get_pull_target() ~= nil
+            or self:get_party():get_player():get_status() == 'Engaged' or self:get_camp_position() == nil then
         return false
     end
 
@@ -398,7 +351,7 @@ function Puller:return_to_camp()
     end
 
     if distance > 2 then
-        local return_to_camp_action = RunToLocationAction.new(self:get_camp_position()[1], self:get_camp_position()[2], self:get_camp_position()[3], 2.0, "Return to camp")
+        local return_to_camp_action = RunToLocationAction.new(self:get_camp_position()[1], self:get_camp_position()[2], self:get_camp_position()[3], 2.0)
         return_to_camp_action.identifier = "Return to camp"
 
         self.action_queue:push_action(return_to_camp_action, true)
